Huge Improvements to stability and app function and other updates
The app doesnt use the chaquopy library no more cause it was causing many problems and it was slow, also i rewrote many parts of the code to minimize on api calls and improve performance. I used the retrofit library for all my api calls simplifing the process. Also i used the mp3agic library to update the metadata on the downloaded songs instead of the mutagen library which was used with chaquopy. Last but not least i redisgned the song download tab with an new download screen